Move Op and OpProof into dedicated modules.
